Move draw map UI components to a dedicated folder
There are going to be a few more components to add to the draw map page.
It would be nice to group these together to reduce the number of files
in the components folder and reduce the size and number of imports in
individual files.
